Rename EMPTY_SECRET into ABSENT_SECRET
All other constants that are named EMPTY_* have non-zero length and
have zeros in it. I also want an EMPTY_SECRET that has 32 bytes of
zeros.

So I rename EMPTY_SECRET = b"" into ABSENT_SECRET.  This value is
currently being used to indicate the absense of the secret.

https://github.com/raiden-network/raiden/issues/3944 keeps track of
replacing ABSENT_SECRET with None.