Stricter rules for pending greenlets on tests
Some of the tests are currently flaky, the test_api_withdraw is a good
example. The logs show that the test finished executing successfully,
but got stuck during the fixture cleanup.

It is not clear at which point during the fixture cleanup the test is
hanging, however it is after the call to `app.stop()` (There are logs
for sucessfull stop of the apps).

This adds an extra pylint rule ensure that the `killall` is not the
culprit, and also makes the rule for cleaning up stricter (instead of
just killing any pending greenlets, now the test will also fail).