Remove withdraw / deposit locks.
Withdraw lock is removed since it does not perform any token-specific
transactions and therefore only the channel-lock was required.

Deposit lock was removed because the lock was introduced in the token
proxy which would prevent concurrent approve / transfer transactions
from being performed. Again, in this case the token lock solves this issue.