fixed tests
Parity seems to have a more conservative gas estimation for deploying
smart contracts. Because of that the initial gas constant had a larger
value. The problem is that the constant was large enough for the
deployment and the transaction execution.

So, this commit merges the geth and parity tests, and instead of
estimating the gas usage before hand, it just burns the extra gas before
sending the transaction.