architecture: split blockchain event handling
This splits the handling of blockchain events into multiple modules:

- The module blockchain.state is responsible to enrich the blockchain
  event data with data from the database, queries from the smart
  contract, etc.
- The module blockchain.decode is exclusively responsible to translate a
  blockchain event into a state change, any additional data that
  requires IO is retrieved through blockchain.state
- The module blockchain_events_handler now is exclusively for
  non-protocol related logic. It handles auxiliary services like the PFS
  and the connection manager.

These changes were introduced to improve decoupling and to allow all the
events for a given block to be processed in a single transaction. The
processing of all events on a single transaction is necessary to fix
race conditions and expose a consistent state view to the any REST API
client.