Fixed race condition on test_send_queued
When a node is sending a transfer forward is does a connectivety check
for each available route, and then offline partners are then filtered
out. The test had a flaky condition that would cause the presence
change for `app1.stop()` to be processed, which resulted in an empty
transfer queue, leading to a failure on the assert.