Remove signature checks from `_data_to_sign`
The check for `reward_proof_signature` does make any sense, because
we're just gathering the data to create that signature in that function.

I've removed the check for `non_closing_signature` along with it because
having it in this place seemed inconsistent (we have no such checks in
`_data_to_sign` for any message) and I don't think this is a good place
for such checks.

Before this PR, the check was applied on serialization. I don't see a
place where I could hook into to reproduce that behaviour and it
probably makes more sense to move towards signed-by-design messages,
anyway.