New directories for consecutive test runs
Tests marked as flaky execute multiple times with the same parametrize
arguments. Will leads to a conflict with some of the file and
directories names that causes the second run to fail.

This fixes that problem by incrementally trying a few new characters at
the end of the storage path.