Remove redundant fields from `message_hash`
In https://github.com/raiden-network/raiden/pull/4284 I tried to keep
all signatures identical. Now it's obvious that all `EnvelopeMessage`
subclasses have a signature that is based on a BP with all additional
fields put into the `additional_hash`. Due to this construction, the
`message_hash` which is used to calculate the `additional_hash` can be
reduced to all fields that are not already included in the BP.

The padding is also unnecessary and was only included to preserve the
old behaviour during the large refactoring PR mentioned above. It's
usually a good idea to separate refactoring and behavior changes.