Fixed DoS attack vector
Messages are being decoded using marshmallow, which maps the json to a
message class based on the `_type` field of the message. If any of the
constructor arguments are missing, marshmallow will raise a Validation
error, which was not being properly handled, and it would lead to a
crash of the `_handle_message` and subsequently the node.