Fixed race condition on test with entire deposit
The test was not waiting for the mediator to synchronize with the first
transfer, so this was possible:

- App0 sends LockedTransfer to App1, which forwards it to App2
- App2 requests the secret and reveals it backwards
- App0 and App1 send the unlock (but none receive yet)
- App2 receives the unlock

At the point above, from App0's perspective the transfer is completed,
the unlock has been sent. Same thing for the App2, the Unlock has been
received, however that is not the same thing for App1.

If the second transfer is started fast enough, it is possible that App1
will reject the payment because the channel App1->App0 doesn't yet have
enough capacity (the 200 tokens are still locked from App1's
perspective).