Reverted pytest-timeout timeout_func_only
That configuration flag was enabled because the fixture teardown for the
parity private chain was hanging. The hanging caused timeouts for some
tests, which made the build flaky. Because of this the flag was added,
making the test fail only if running it took too long.

The problem with the above approach is that it does not remove the
flakyness of the fixture. The fixture teardown may still hang, and
without the timeout exception it will be killed by the continious
integrations timeout, which does not produce logs, making it hard to
debug the issue.