Increase timeout for CLI tests.
Without the flag timeout_func_only the time it takes to setup and
teardown the fixtures must be taken into account. The test
test_cli_wrong_rpc_endpoint was used as a baseline, since it failed with
a 45s timeout, another random successful test was picked and it needed
48s to completely run (including fixture setup/teardown). Because of
this the timeout was increased to 120s.