Removes a corner were a timeout kills the test suite.
`pytest-timeout` uses the `pytest_runtest_protocol` hook to install and
clear the timeout handlers. The problem is that this hook does not
handle exceptions, there are a few places were an exception can be
raised and result in a build failure because pytest itself is killed.

This change makes the period of time were a timeout can be raised
shorter. This is achieve by installing the signal inside
`pytest_runtest_setup`, and uinstalling it at `pytest_runtest_teardown`.

Ideally an exception should only be raised while execution is inside
`pytest_runtest_setup`, `pytest_runtest_call` or
`pytest_runtest_teardown`, but **not** in-between these calls. However
there is no hook available for that and this change would require
additional complex logic.

This means that even with this change, there is still the possibility of
a timeout raising an exception and resulting in an internal error.

This change also prints the stack traces of the greenlets, and not just
the python threads, this can be useful to find deadlocks.