Added tool to kill dormant processes
This tool is being added to help misbehaving builds. The continuous
integration services, e.g. Travis and CircleCI, will kill the build if
the process doesn't output any data. Because the whole environment is
kill, e.g. the container or the VM is killed, there is no logs available
to debug the problem.

This tool should be used to send a user signal before the CI timeout is
reached, the test suite should run a signal handler to print debugging
data and exit. If the test suite doesn't exit in a timely manner then it
will be forcefully killed.