Add a check in the tests that the geth version is supported
I was just beaten by thinking the flaky tests were due to some problem
that ended up being due to geth v1.9.0. So I am adding this check on
tests to make sure that when we run the tests they are not with a
version that is not supported.

Not sure how to keep config.yaml with the raiden/constants.py