Fix corner cases for fixture timeout
The previous implementation had a corner case that allowed the whole
pytest process to be killed. That happened when an exception was raised
while the pytest's runtestprotocol function was executing, but not
inside a `CallInfo.from_call`, which is the piece of code responsible to
handle exceptions from the setup/call/teardown phases.

To fix the above, this uses the hooks pytest_runtest_setup,
pytest_runtest_call, and pytest_runtest_teardown, to enable and disable
the timeout logic. This only works because these three hooks are always
called from inside a `CallInfo.from_call`.

The previous implementation was flawed because the timeout was enabled
between the calls. E.g. that allowed the timeout exception to be raised
after the pytest_runtest_call but before pytest_runtest_teardown, which
would raise an exception in an unprotected region.