Stricter rules for timeouts
The teardown must be allowed to run, otherwise the fixtures cleanup
cannot be executed. Because of this the timeout for the teardown must be
different than the timeout for the setup and call phases of the test
execution.

This commit splits the previous timeout setting into two new settings.
One just for the setup+call phase, and another for the teardown.

Additionally, this makes the timeout settings required. This is
necessary because these timeouts are not used only to detect misbehaving
tests. Timeouts are also used to ensure that we will get a useful report
from the continuous integration. Because of the way the CI works, if a
build doesn't output any data for a given period the whole build is
killed, this means a test cannot take longer than that setting to run.
The timeout is used to enforce that boundary.