Timely cleanup
Until now, the managers timeout were completely unrelated to the test's
timeout, which led to hard to debug flakiness. It happened that the test
setup and execution used most of allocated time, once the teardown
process started, the amount of time left was minimal. Because the
fixture used an arbitrary value, completely unaware of how much time was
left, the teardown timedout too often.

The above can be described with the following diagram:

- timeout    |-----------------|
- setup      |------|
- call                |----|
- teardown                  |-=====|

Once the teardown was called, very little time was left. The teardown
unware of that used large timeouts, which led to frequent failures
during teardown.

This builds on the previous commits, were a teardown timeout is
separated from the setup and call timeouts. Giving a predicatable amount
of time for the whole teardown process. On top of that, the existing
processes that are variable in time have theirs allocated times
drasticaly reduced. The hope is that all the teardowns will be sucessful
before the teardown happens.