Fixes race condition from #4344
This is a tricky race condition, it is unlikely to happen with the
current version of the code, but it is fixed by this commit neverthless.
The race condition manifests itself when the node is processing an event
from a pruned block, and it needs to fetch additional data from the
blockchain. The problem was that the existing code used `latest` by
default, which effectively fetched unconfirmed data and used it as if it
was confirmed.

Consider the following:

- Alice opens a channel with Bob
- Bob crashes
- Bob comes back online after at least `pruned_block`s.
- At this point, Bob has to synchronize with the blockchain, it starts
  polling for the missing events, once it processes the ChannelOpen
  event, it can fetch additinal data. The existing code is fetching the
  channel deposit from the blockchain.
- If Alice sends a transaction around this time, and Bob uses the block
  `latest` in the pruned block handler, the unconfirmed deposit of Alice
  will be used.

To prevent the above two things should be done:

- Fetching additional data from the blockchain on a event mapping should
  be kept to a minimum.
- If fetching additional data is necessary, then it must use a confirmed
  block instead of latest.