Unskip cli tests
The tests `test_cli_full_init_dev` and
`test_cli_manual_account_selection` failed because the existing timeouts
were too short. This is a problem of composability of timeouts, were the
outter timeout can be set without synchronizing with the inner timeouts.

What happened with the above mentioned tests is that the teardown was
expecting to have 10 seconds of time to properly cleanup the private
chain, while the complete test had only 45 seconds in total, and the
setup and test ran took 35s out of the 45s. This left the test with a
race condition were sometimes it was enough time to do a teardown, and
sometimes it was not (If this was enough time depended on the actual
time the test took to run, or if the parity process was well behaved)

Some additional edge cases existed in the timeout setup that were fixed
by #4458. These tests should now run sucessfully with 45 seconds.