Fix for issue #3915 - refunds for initiators were not leading to re-route of transfers
The fix involves splitting existing ReceiveTransferRefundCancelRoute into two different
state changes: `ReceiveTransferCancelRoute` and `ActionTransferReroute`. The first is
always handled, while the re-routing is only done for transfers that are not atomic swaps.

The re-route handling now works by creating a copy of the existing initiator task (keyed
by the new secrethash) which can then go through the rest of the flow, while the original
route is left there until its lock is expired.