Add typing_extensions to production requirements
This is done only because of the Literal["foo", "boo"] construct. It
needs the typing extensions in order to exist.

Another way to go around adding this to requirements is to define all
literals in `raiden/utils/typing.py` and with an `if TYPE_CHECKING:`
change the type to `Any`.