Remove deposit event confirmation handling
The code was written when the incoming event was unconfirmed. Now, all
events are already confirmed when an Event instance is created. The
confirmation handling that is left is not only unnecessary, but even
creates wrong behavior by adding an additional latency of
`confirmation_blocks` to the event processing.

This commit removes the confirmation handling and the respective tests.
No other changes intended.

Closes https://github.com/raiden-network/raiden/issues/4377.