Fixes invalid block number on state retrieval
To create the state changes ContractReceiveChannelSettled and
ContractReceiveChannelNew fetching additional data from the blockchain
is necessary.

The data used for these state changes has to be both:

- Confirmed
- Available in a block that has not yet been pruned.

The first point was a source of race conditions that led to flakyness on
some tests, were it was possible for an unconfirmed deposit to be seen
by the node, leading to subtle errors and being a possible attack
vector.

The second point was another source of flakyness, if the node happened
to go offline for a while, the confirmed block in the node's state was
used to guarantee the fetched data itself was confirmed, however on
restarts this block could be too old, therefore it could be a pruned
block.

To fix both of the problems above, the decoding function now receives
the latest_confirmed_block, which is computed by the RaidenService based
on the latest block number. The assumption here is that an ethereum node
can prune any block, however it will always have a confirmed block
available.