Replace `SuccessOrError` type alias for a proper class.
We are using currently a custom-defined `SuccessOrError` type for the
cases where we need to test if a condition is valid *and* we also would
like to have a message indicating *why* the condition is not true.

This `SuccessOrError` is was basically an alias to a `Tuple[bool,
str]`, which did not help much in terms of programmer ergonomics (we
would still have to destructure the tuple to check on the condition)
and did not help much with type-enforcing.

The alternative proposed by #4926 was to replace `SuccessOrError` with
a simple `ErrorMsg` type and having the functions returning
`Optional[ErrorMsg]`.  This could make things a bit simpler but still
did not help much in the ergonomics part (e.g, the return value would
be `True` whenever a function had errors and the condition being
tested was `False`.)

The solution proposed and implemented here transforms SuccessOrError
into a proper class. Objects of `SuccessOrError` can be instantiated
with multiple error messages. The object itself has its "truthyness"
determined by the presence or absence or error messages, i.e, a object
without any error messages is `True`. This allows us to write functions
that return `SuccessOrError` objects and use the comparison right away
if we want or check for the errors in case the condition fails.