Allow running tests on geth < 1.9
Versions >= 1.9 require `--allow-insecure-unlock` while versions before
can't take this parameter. To make the tests run on either version, we
check for the version and only add the parameter for later geth
versions.

As long as we officially support geth 1.8 it is good to have a way of
running the tests on it.