Fixed fixture setup
The function wait_for_channels was only waiting for one of the sides to
see the deposits. This happened because the argument names for
wait_for_usable_channel were confusing, and instead of the function
being called twice with the `app0` and `app1` arguments swapped around,
it was called twice with the same arguments.