Fixed assert_balance
The previous version of the code was using a check like:

    balance0 > balance1

To detect divergence in the deposits. This does work only for the
scenario where app0 has seen the deposit before app1. But it does not
work in the other way, when app1 has seen a deposit before app0. Which
is an important synchronization problem that this function should
detect.

The fix is to compare against inequality. So if any of the nodes has
seen a deposit that the other has not yet, the assertion will fail. This
does not catch bugs were *both* nodes ignore or miss an event, but
without extra information the assert function cannot check that.

The function still only asserts on the deposit of a single node, calling
it twice checks the synchronization in the other way around.