Fix test_regression_transport_global_queues_are_initialized_on_restart_for_services
Tests seems fine in general, however it was never adapted to the
addition of the PFSFeeUpdate message. This was added and a more fine
grained checking for which messages should be queued added.

One problem than was found after this is that two identical
PFSCapacityUpdate were queued. This was fixed by calling
update_monitoring_service_from_balance_proof directly.