Removed extranoeous RPC call to load the balances
The Raiden node will synchronize with the blockchain by fetching all the
events emited by the smart contract. This means that if a deposit is
done, eventually the node will fetch the corresponding channel deposit
event and update the off-chain state.

It is every important for these updates to be done on top of a confirmed
block, which is currently properly handled by the alarm task. The
existing code was fetching additional data from the blockchain, which
was a simple source of bugs.

This just remove the unecessary extra RPC call, and relies exclusively
on event fetching to update the channel's balance. It removes the extra
RPC call and should make for an easier to understand data flow.