Fixed the case were there are not pending locks
When all the locks have been removed, the nodes will use the keccak hash
of the empty string, `keccak(b"")`, which is set in the constant
LOCKSROOT_OF_NO_LOCKS.

When no balance proof is provided, the node has to call settle with an
emtpy balance proof, which would use the null hash `b"\x00" * 32`.

Therefore both of the values above have to be handled and accepted as
the zero value.