Removed sleeps from `test_matrix_transport.py`
The new code is not very beautiful, but it should be more robust than
the sleeps. Suggestions on how to write it in a nicer way are welcome!

The specific timeouts are a bit arbitrary, but since we won't wait for
the timeout unless our test it broken, it's OK to have the timeouts a
bit higher.