Regression test for 4470
Test that opening a channel via the API provides the confirmed block and not
the latest block as given_block. The discrepancy there lead to potential timing issues where
the token network was deployed for the state in the "latest" block but not yet
in the confirmed state and a BadFunctionCallOutput exception was
thrown from web3.

Regression test for #4470

At the moment of this commit where the fix is not yet in the test
fails with:

```
 web3.exceptions.BadFunctionCallOutput: Could not decode contract function call getChannelIdentifier return data b'' for output_types ['uint256']
```