Fixed test_detect_balance_proof_change
The test was instantiating invalid objects, this rewrites the
instantiation and tries to instantiate only valid objects. The approach
takes was simple:

Instantiated the simplest objects first, and gradually add new state,
when a new state is created it is compared against the previous states,
and if a new balance proof was introduced it is asserted accordingly.