Treat trying to deposit on a closed channel as recoverable error
Maybe fix #4451

If a channel is not open, or is in a closing state (still open but
waiting on the closing transaction result) the `set_total_deposit`
function was throwing a `ValueError`.

That is a valid and recoverable race condition so this commit
introduces an appropriate error and handles it in all places where it
can be thrown.