Check `settle_timeout` threshold for success, too
In https://github.com/raiden-network/raiden/pull/4766 the question comes
up whether the contracts check the settle timeout in exactly the same
way. We've already checked for failure of values outside the expected
bounds, but to make sure things match in all cases, we also have to test
for success.