Improved messages for failure on deposit
This reuses the failure messages from the gas estimation in the
transaction failure for a the channel deposit. This change is only
cosmetic, to improve the error message.

This commit also removes _get_channel_state, that was only adding an
extra indirection layer without any useful logic, it made bit harder to
review the code.