Added check for the channel identifier.
The function `set_total_deposit` is provided with a channel identifier
and a partner address. This is necessary because the smart contract does
not store the participants addresses to save storage and reduces costs.

These values must validated, otherwise it is possible to call the
proxy's `set_total_deposit` with a correct channel id which the current
node is not a participant, where correct is just a channel that was
created at some point in the past.

The check is not complete: Because of state prunning the preconditions
cannot always be checked, and because once theis removed the id is
removed from the smart contract storage, the check can only be performed
if the state of the channel is at OPENED or CLOSED.