Improve pre-commit config
This adds mypy to the pre-commit config and changes the way pylint is
invoked via pre-commit.

Previously a repo-local pylint hook was used. That lead to pylint only
being invokable if the project venv was activated in the context the
pre-commit hook was called from. That usually isn't the case when using
GUI git clients or IDE / editor integrations.

Instead it uses the default pylint mirror provided by pre-commit with a
wrapper script that ensures the project requirements are available and
adds the project root to the `PYTHONPATH`. That
allows the project local pylint plugins to be importable while still
using the normal pre-commit hook venv isolation.

The same wrapper is used for mypy since it also needs access to third
party library code in order to successfully check type definitions.

Also adds some other useful sanity check hooks.

Small drive-by fix:
- Fix the pylint gevent_checker plugin to not choke on uninferrable values