Remove --nodiscovery from the fixture setup
`--nodiscovery` was given to make sure the test network didn't try to
join any mainnet nodes. This however has the negative side effect of
disabling discovery even if the `--bootnodes` flag is given, meaning
that only the miner node can be used for an integration test.

To allow more than one geth node on a test network that flag cannot be
used. Our private network should be independent of the main network
because we overwrite the `--bootnodes`.