token_network_register fetches the chain_state
for checking preconditions.

This commit is just a refactoring towards adding
precondition checking in the TokenNetworkRegistry proxy.

The proxy needs a blockhash used for condition checks in
the API.

This commit changes get_tokens_list() so that it takes a
ChannelState.  Also token_network_register() now passes
a channel state to this function.