Remove add_token_without_limits() b/c never used
This function is useful for interacting with TokenNetworkRegistry
of RedEyes and similar versions, but currently it is never used.

Moreover, when I try to change the interface of _add_token()
function, I need to make a spurious change in add_token_without_limits().
So it is about time that this function should be removed from the codebase.