Reduce the number of state changes for fee updates
In https://github.com/raiden-network/raiden/pull/4827, we've added
`TriggerFeeUpdate` in addition to having `ActionChannelUpdateFee`. This
commit remove these in favor of one simple `SendPFSFeeUpdate` event. The
main purpose of this is to make sure that balances and fees are always
consistent by changing them during the same state change.

Pro:
- Easier to enforce consistency
- No `ActionChannelUpdateFee` in `blockchain_events_handler.py` (I don't
  think it belonged there)
- Less state changes
- Shorter code

Cons:
- Needed to add `fee_config` to deposit and withdraw state changes
- Will probably need to reintroduce a state change once users can change
  fees during runtime.

Closes https://github.com/raiden-network/raiden/issues/4852.