Enable auto-assign probot
Configures https://probot.github.io/apps/auto-assign/ to add one
reviewer to each new PR in the Raiden repo.

I've requested the App above to be added to the Raiden repo. An admin
will have to accept that request before it starts working.

I've not tested this configuration. If it does not work as intended,
I'll create a repo to test this.

Closes https://github.com/raiden-network/team/issues/602

[skip tests]