Added metadata
Added metadata classes for the token network, token network registry,
and blockchain service.

Because the token network is not deployed by our scripts, there is no
metadata available for these smart contracts, so the a few of the field
in the metadata class have to be optional.

Ideally the code should be refactored to always use the
BlockchainService.token_network_from_registry method since that will
fetch the missing metadata. This is important to reduce the range of
blocks used by the filters of a given token network.