Skip redundant `WithdrawExpired` signature check
The signature is already checked when the `ReceiveWithdrawExpired` is
created. The `sender` used when creating the instance is recovered from
the signature. So all we have to to is to check that this `sender` is
the right one. This is also done in `is_valid_withdraw_expired`.

My main reason for removing this redundant check is that it expects the
`WithdrawExpired` to have the same signature as the other withdraw
messages and I want to change this as part of
https://github.com/raiden-network/raiden/issues/4934.