tightened type of StatelessFilter._do_get_new_entries
Now the to_block is always a BlockNumber, this means that
`block_specification_to_number` doesn't need to be called anymore. And
since the function is not being used anymore, it was removed.

This change fixed a race condition, because
`block_specification_to_number` could request the block number for
`latest`, it was possible that its value changed meanwhile,
de-synchronizing the values.