Mention situations in RaidenValidationError
because before this commit, the description read like this exception
is just about the validation of values.  Actually, this exception
can be used to validate values with respect to the current state of
the client or the blockchain.  For example, even a valid
address can cause RaidenValiationError if the address is supposed to
be registered but it's already been registered.  So the validations
are allowed to take into account things other than the value itself,
and still raise RaidenValidationError.

This commit comes from the discussion here
https://github.com/raiden-network/raiden/issues/4878#issuecomment-534095416