Batch event polling
This introduces batch event polling, the goal is to collapse all
requests into one, so that instead of having one request per filter,
only a single batch request is done for all filters.

This is particularly important for our test enviornments, since it is
possible for a token network registry to end up with hundreds of
registered tokens, were previously that was an equal amount of JSON-RPC
requests per block.

Additionally, this fixes the bug #4498 for the runtime of the node too,
and not just for the initialization. This is achieve by only returning a
batch of events after all filters have been installed and fetched.