Set reveal timeout after the channel is created.
This also removes a condition from the constructor of the channel
state which checks that settle timeout is larger than the reveal
timeout.
At the time of the channel being created, Raiden receives the new
channel created event with the settle timeout that was provided to
channel at the time of creation. However, reveal timeout in blockchain
event handling is set to the default Raiden value (which could be
larger than the provided settle timeout). This will crash the Raiden
node. Therefore, it was decided to remove this check and basically
accept the new channel while making it unusable for mediating
transfers in case the settle timeout is smaller than the reveal
timeout.