handle all status codes in resolver client
This handles all status codes in the resolver client logic. Previously,
only codes 200 and 404 were checked, if we received any others then the
response would be ignored and the loop to retry the resolver endpoint
would continue.

This treats the 200 status as a success that breaks the loop and
attempts to use the secret returned by the response, and 503 `SERVICE
UNAVAILABLE` as an indicator that we should attempt the request again.
Any other code is regarded as a failure and makes the method return
`False`.