Add dockerhub build hook
This should fix #5049

In order to build the proper raiden version from `Dockerfile`, we need
to pass the intended `RAIDENVERSION` build argument. According to
https://github.com/docker/hub-feedback/issues/508 the current way of
doing this is via build hook.
dockerhub has the `SOURCE_BRANCH` environment variable available, that
should translate just fine to the values we want in `RAIDENVERSION`.