Add unit test for resuming waiting transfers
This adds a test that asserts on the resume logic for a transfer, that
is received "too early", i.e. at a time, where the lock expiration is at
a block number higher than the current block number + settlement
timeout.

The transfer is in `"waiting"` state and will only be forwarded, once
our view of the blockchain is at a block number safe for forwarding.