Renamed CANONICAL_IDENTIFIER_GLOBAL_QUEUE
The function `enqueue_global` could be a bit confusing since there are
both global queues and global rooms. The function `enqueue_global` is
used to put a `Delivered` message into the global queue (which is not
sent to the global room).

To avoid the above confusion this changes from "global queue" to
"unordered queue".

The unordered queue is used to send messages that don't need to be
processed in a given order to the partner node. Which is different from
a global room, which is used to broadcast messages to multiple nodes
(particularly the services).