Simplified registration code
The registration code had a hard coded number of retries, were each
retry used a new userid for registration. This code had a few issues:

- The code was introduced to prevent a DoS attack, were an attacker
would discover the user's username from previous iterations, and if the
matrix federation changed, the attacker would register these usernames
before the user. It is not know how difficult is for someone to peform
this attack, however since there was code in place to protected against
it, I assume it is possible to happen. This change the code so that it
is not possible for the attack happen after a bound number of iteration,
by completely removing the usage of the seed for the PRNG.
- Because of the above change, a loop is not necessary anymore, making
the code slightly shorter and easier to understand.
- Because now a registartion will always use a new userid, it is not
necessary anymore to try to login to one of the existing accounts.