Use Fractions to avoid inaccuracy for big ints
The standard double-precision floats can't precisely represent integers
larger than ~10**16. This has the effect that balances above that caused
relatively large inaccuracies even for small payments (errors of
multiple tokens for a single token transfer).

This is fixed by calculating with fractions instead of floats at the
cost of decreased performance. Benchmarks will follow to assess the
impact. This might especially problematic for the PFS when optimizing
routes.