New test: adding fee is inverse of removing fee
The PFS starts with the target amount and adds fees while the mediators
start with the transfer amount and deduct fees. These operations must be
the exact inverse.

This new test produces easier to debug test failures than
`test_fee_round_trip`, which makes it a useful test to have.