Increase pfs request timeout from 1 to 10 seconds
Fix #5078

When  many requests are sent to the PFS at the same time and the pathfinding timeout is only 1 second

https://github.com/raiden-network/raiden/blob/1ff6e6e22017b16a22fcef0e23535d5c41c617cc/raiden/constants.py#L153

then one of the requests does not make it within this one second window and is killed by the node before even reaching the PFS. Causing it to revert back to internal routing as explained [here](https://github.com/raiden-network/raiden/issues/5078#issuecomment-548009951).

Increasing the request timeout fixes this.