Removed random user id / Assume Matrix auth plugin
Random user IDs with the random sufix are not accepted by the current
version of the `eth_auth_provider`, any attempt to use that user name
would result in a crash. Therefore its usage was removed.

This changes the login process to assume the server will always have our
authentication plugin installed as requested by @ulope. The assumption
here is the Matrix servers used with Raiden are not part of the normal
Matrix federation, otheriwse these servers would replicate normal chat
messages that are not of interest to the Raiden network. Since we assume
there is this initial configuration step, it is also assume the plugin
will be installed.