Cosmetic changes
This is not adding any feature or fixing any bugs, it is just making
sure the cached user's displayname is warm before it is returned from
the `_get_user` function. It was just surprising to me it was not
warmed, but the previous code would eventually result in a warm
displayname since the `User` instance was shared.