`_get_room_ids_for_address` changes
This commits does:

- Removes the automatic upgrade for the value of `network.raiden.rooms`
to a list. Since Red Eyes this value is always a list of strings, the
old format where it contains just a single room_id doesn't have to be
supported anymore.
- Adds a check for every entry of the metadata. If the values are
corrupted either there is a Raiden bug or a malicious matrix server.
- Reduces the critical section of the self._account_data_lock to exactly
the lines that interact with the `client.account_data`.