Removed `DummyUser`
`DummyUser` was used exclusively because of the `__hash__` method, which
is not defined in the `User` class from the Matrix SDK. This was
necessary to check the values in the presence storage, which was a
mapping from `User` => `Presence`.

The problem is that for this to work the `DummyUser` instance had to be
used in the presence handler, meaning the `cached_user` could not be
used.

This commit removes the `DummyUser` class and relies exclusively on the
`User` from the Matrix SDK, which should simplify the test. And to fix
the problem with the lookup it only changed the mapping from `User` =>
`Presence` to a mapping of `user_id` => `Presence`, where `user_id` is a
plain old `str` that is stable for hashing.