Fix runtime error: dict modified while iterating
The container `_address_to_userids` is a defaultdict used to store the
known `UserId`s for a given `Address`. This container can be modified by
multiple threads (e.g. `UserAddressManager`, and blockchain event
handling that may start a whitelist/healthchecking), this can lead to
the container being modified while it is being iterated, which is a
runtime error.

This worksaround the problem by iterating over a copy of the keys,
instead of iterating over the dictionary key view.