Added explicit cache management for display name
The previous implementation relied on the Matrix SDK to keep a cache of
`User` instances withe a warm `displayname`. The problem is that the
cache could do one of the two:

- Polute the SDK's Room instance with users that were not part of the
room
- Only partially cover the `User`s, since any user that was not in the
discovery room should not be part of the cache to avoid polluting the
room.

To fix the above, instead of using a `Room`'s private attribute as a
cache, an explicitand simple  cache is introduced that does not have the
aforementioned problems.